api/requirements.txt
fastapi
uvicorn[standard]
pydantic
minio
requests
python-multipart
python-dotenv
litellm
PyPDF2
selenium
webdriver-manager
html2text
fastmcp
git+https://github.com/EvoAgentX/EvoAgentX.git